requirements.txt
click==7.1.1
docopt==0.6.2
Flask==1.1.2
gunicorn==19.10.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.16.6
opencv-contrib-python==4.2.0.32
opencv-python==4.2.0.32
stomp.py==4.1.23
Werkzeug==1.0.1
